% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_summary.R
\name{fishsummary}
\alias{fishsummary}
\title{Summarize a set of records downloaded from FishNet2}
\usage{
fishsummary(input, verbose = TRUE)
}
\arguments{
\item{input}{A dataframe in FishNet2 standard format (by using read.csv())}

\item{verbose}{Print progress and information messages. Default: TRUE}
}
\value{
A list of summary statistics

# summarize occurrence records
}
\description{
Creates a simple summary of data returned by a FishNet2 search.
}
