% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_tissue.R
\name{has_tissue}
\alias{has_tissue}
\title{Filter a set of records downloaded from FishNet2 by Tissue column}
\usage{
has_tissue(input, verbose = TRUE)
}
\arguments{
\item{input}{A dataframe in FishNet2 standard format (by using read.csv())}

\item{verbose}{Print progress and information messages. Default: TRUE}
}
\value{
Filtered dataset with records that do not have a blank tissue value

# summarize occurrence records
}
\description{
Filters data returned by a FishNet2 search for records that include tissue information.
}
\examples{
has_tissue(louisiana,TRUE)

}
