% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ictaluridae}
\alias{ictaluridae}
\title{Dataset of Ictaluridae Taxon from Years 2017 to 2019}
\format{
A data frame with 273 rows and 16 variables: \describe{
  \item{InstitutionCode}{unique code given to institution who owns the data}
  \item{IndividualCount}{Number of fish individuals}
  \item{ScientificName}{Scientific name of fish observation}
  \item{Family}{Family of fish observation}
  \item{PreparationType}{Type of preparation}
  \item{Tissues}{Whether observation contains tissues}
  \item{Latitude}{Latitude observed}
  \item{Longitude}{Longitude observed}
  \item{Country}{Country that lot was observed in}
  \item{StateProvince}{State or province where lot was observed}
  \item{County}{County that lot was observed in}
  \item{YearCollected}{Year collected}
  \item{MonthCollected}{Month collected}
  \item{DayCollected}{Day collected}
  \item{BasisOfRecord}{Preserved Specimen}
  \item{DateLastModified}{Data record last modified in database}

}
}
\source{
\url{http://www.fishnet2.net/search.aspx?t=ictaluridae&d=2017-2019}
}
\usage{
ictaluridae
}
\description{
A dataset as a result of a search query of taxon 'Ictaluridae' and date range
'2017-2019' on fishnet2.net
}
\keyword{datasets}
