% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{louisiana}
\alias{louisiana}
\title{Dataset of Records from Louisiana from Years 2005 to 2006}
\format{
A data frame with 273 rows and 20 variables: \describe{
  \item{InstitutionCode}{unique code given to institution who owns the data}
  \item{CollectionCode}{Collection Code}
  \item{IndividualCount}{Number of fish individuals}
  \item{ScientificName}{Scientific name of fish observation}
  \item{Family}{Family of fish observation}
  \item{PreparationType}{Type of preparation}
  \item{Tissues}{Whether observation contains tissues}
  \item{Latitude}{Latitude observed}
  \item{Longitude}{Longitude observed}
  \item{HorizontalDatum}{Horizontal Datum}
  \item{Country}{Country that lot was observed in}
  \item{StateProvince}{State or province where lot was observed}
  \item{County}{County that lot was observed in}
  \item{YearCollected}{Year collected}
  \item{MonthCollected}{Month collected}
  \item{DayCollected}{Day collected}
  \item{Collector}{Name of collector}
  \item{GeorefMethod}{Geo Reference Method}
  \item{BasisOfRecord}{Preserved Specimen}
  \item{DateLastModified}{Data record last modified in database}

}
}
\source{
\url{http://www.fishnet2.net/search.aspx?l=+Louisiana&d=2005-2006}
}
\usage{
louisiana
}
\description{
A dataset as a result of a search query of location 'Louisiana' and date range
'2005-2006' on fishnet2.net
}
\keyword{datasets}
