% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_map.R
\name{occ_map}
\alias{occ_map}
\title{Plot Longitude and Latitude Points on World Map}
\usage{
occ_map(df, color = "darkred")
}
\arguments{
\item{df}{A dataframe in FishNet2 standard format (by using read.csv())}

\item{color}{Color of plotted points, default is dark red}
}
\value{
Plot of latitude and longitude points on world map
}
\description{
\code{occ_map} returns a plot with columns 'Longitude' and 'Latitude' in FishNet2
dataframe on a world map.
}
\details{
This is a function to get a plot of occurrence records from FishNet2 search query.
Parameter is a dataframe that must have the columns 'Longitude' and 'Latitude'.
NA values are removed in the function.
}
\examples{
occ_map(ictaluridae)
}
