% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_plots.R
\name{top_n_plots}
\alias{top_n_plots}
\title{Outputs a bar graph giving the top n in frequency in specified column of
dataframe}
\usage{
top_n_plots(df, n, colName, color = "default colors")
}
\arguments{
\item{df}{A dataframe in FishNet2 standard format (by using read.csv())}

\item{n}{The number of the labels with the highest frequencies to be included
in the graph}

\item{colName}{The column name that the graph outputs}

\item{color}{Color of the bars, by default is a different color for each bar}
}
\value{
A bar graph
}
\description{
\code{top_n_plots} returns a bar graph that shows the top n (n is given as a
parameter) labels in a given column in the dataframe with the highest
frequency
}
\details{
This is a function to create and output a bar graph giving the top n in
frequency in specified column of dataframe (columns include 'ScientificName',
'Family', 'Country','State/Province','County').
}
\examples{
top_n_plots(ictaluridae,10,"ScientificName")
}
